requirements.txt
psutil
openai
cryptg
uvloop
valkey
orjson
asyncpg
requests
tgcrypto
telethon
google-genai
aiohttp[speedups]
python-telegram-bot
